// drr/src/UI/src/model/forecastStatus.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * DRR API
 * OpenAPI spec version: 1.0.0
 */

export type ForecastStatus = typeof ForecastStatus[keyof typeof ForecastStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ForecastStatus = {
  Pending: 'Pending',
  Review: 'Review',
  Approved: 'Approved',
  Rejected: 'Rejected',
} as const;